+++ lib/magic.dart
@@ -123,6 +123,9 @@ export 'src/database/seeding/seeder.dart';
 export 'src/database/eloquent/model.dart';
 export 'src/database/eloquent/concerns/has_timestamps.dart';
 export 'src/database/eloquent/concerns/interacts_with_persistence.dart';
+export 'src/database/eloquent/casts/casts_attributes.dart';
+export 'src/database/eloquent/casts/enum_cast.dart';
+export 'src/database/eloquent/casts/list_cast.dart';
 
 // Authentication
 export 'src/auth/authenticatable.dart';